refactor: modernize syntax and harden settings load/save failure paths
Adopt primary constructors, collection expressions, and Lock across services and view models. Replace `using` with `await using` for async streams in JsonPidStore. Move local functions after return statements and reformat XAML for readability.

Harden SettingsViewModel to handle load and save failures gracefully disable Save when load fails, show a failure banner on save errors, and keep the dialog open so the banner is visible. Add a design-time constructor with valid defaults. Add tests covering load failure, save exception, and design-time constructor behavior.

Update csproj to set StartupObject and normalize AnyCPU platform output paths to x64.

+++ AutoQAC/Views/Helpers/IntDoubleConverter.cs
@@ -6,7 +6,7 @@ namespace AutoQAC.Views.Helpers;
 public sealed class IntDoubleConverter : IValueConverter
 {
     public object Convert(object value, Type targetType, object parameter, string language) =>
-        value is int intValue ? (double)intValue : 0d;
+        value is int intValue ? intValue : 0d;
 
     public object ConvertBack(object value, Type targetType, object parameter, string language) =>
         value is double doubleValue ? (int)Math.Round(doubleValue) : 0;
